feat(testing): Replace RSpec with Minitest
This commit replaces the RSpec testing framework with Minitest. The
change involves updating the Gemfile and cont.gemspec to remove the
RSpec dependency and add the Minitest dependency. The existing RSpec
tests have been rewritten in Minitest syntax and placed in a new file
cont_test.rb. This change is part of an effort to simplify the testing
setup and use a testing framework that is included in Ruby's standard
library.